requirements.txt
accelerate==0.16.0
backoff==2.2.1
cityscapesscripts==2.2.1
git+https://github.com/openai/CLIP.git
decord==0.6.0
dill==0.3.6
einops==0.6.0
ftfy==6.1.1
h5py==3.8.0
inflect==6.0.2
ipython==8.11.0
ipykernel==6.22.0
jupyter==1.0.0
joblib==1.2.0
kornia==0.6.9
matplotlib==3.6.2
nltk==3.8.1
num2words==0.5.12
numpy==1.23.5
omegaconf==2.3.0
git+https://github.com/openai/openai-python.git
opencv_python_headless==4.5.5.64
pandas==1.5.2
Pillow==9.4.0
prettytable==3.6.0
pycocotools==2.0.6
python_dateutil==2.8.2
PyYAML==6.0
qd==0.8.9
regex==2022.10.31
requests==2.28.1
rich==13.3.2
scipy==1.9.3
setuptools==65.6.3
tensorboardX==2.6
tensorflow==2.11.1
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
git+https://github.com/huggingface/transformers.git
wandb==0.13.9
word2number==1.1
yacs==0.1.8
